Refactored to avoid circular loop and reduced the borrowed code
