+++ pyproject.toml
@@ -17,6 +17,7 @@ classifiers = [
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
@@ -31,10 +32,9 @@ classifiers = [
 "Changelog" = "https://github.com/seapagan/simple-toml-settings/blob/main/CHANGELOG.md"
 
 [tool.poetry.dependencies]
-python = ">=3.10,<4.0"
+python = ">=3.9,<4.0"
 rtoml = ">=0.9.0,<1.0.0"
 github-changelog-md = ">=0.9.0"
-xdg-base-dirs = "^6.0.1"
 
 [build-system]
 requires = ["poetry-core"]
@@ -46,7 +46,7 @@ build-backend = "poetry.core.masonry.api"
 mypy = "^1.4.1"
 pre-commit = "^3.3.3"
 pymarkdownlnt = "^0.9.12"
-ruff = ">=0.5.1,<0.6.0"
+ruff = ">=0.1.3,<0.5.0"
 
 # task runner
 poethepoet = ">=0.24.1,<0.27.0"
@@ -60,7 +60,7 @@ pytest-randomly = "^3.13.0"
 pytest-reverse = "^1.7.0"
 pytest-mock = "^3.11.1"
 mock = "^5.1.0"
-faker = ">=19.2,<27.0"
+faker = ">=19.2,<26.0"
 pyfakefs = "^5.2.4"
 
 # documentation
