+++ requirements.txt
@@ -1,54 +1,24 @@
-<<<<<<< HEAD
-certifi==2024.7.4 ; python_version >= "3.10" and python_version < "4.0"
-cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
-charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
-click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
-colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
-cryptography==42.0.8 ; python_version >= "3.10" and python_version < "4.0"
-deprecated==1.2.14 ; python_version >= "3.10" and python_version < "4.0"
-github-changelog-md==0.9.4 ; python_version >= "3.10" and python_version < "4.0"
-idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
-markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
-mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
-pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
-pygithub==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
-pygments==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
-pyjwt[crypto]==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
-pynacl==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
-requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
-rich==13.7.1 ; python_version >= "3.10" and python_version < "4.0"
-rtoml==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
-shellingham==1.5.4 ; python_version >= "3.10" and python_version < "4.0"
-typer==0.12.3 ; python_version >= "3.10" and python_version < "4.0"
-typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
-urllib3==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
-wrapt==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
-xdg-base-dirs==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
-=======
-annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
 certifi==2024.7.4 ; python_version >= "3.9" and python_version < "4.0"
 cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
 charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
 click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
 colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
-cryptography==42.0.5 ; python_version >= "3.9" and python_version < "4.0"
+cryptography==43.0.0 ; python_version >= "3.9" and python_version < "4.0"
 deprecated==1.2.14 ; python_version >= "3.9" and python_version < "4.0"
-github-changelog-md==0.9.3 ; python_version >= "3.9" and python_version < "4.0"
+github-changelog-md==0.9.4 ; python_version >= "3.9" and python_version < "4.0"
 idna==3.7 ; python_version >= "3.9" and python_version < "4.0"
 markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
 mdurl==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
-pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
-pydantic-core==2.16.3 ; python_version >= "3.9" and python_version < "4.0"
-pydantic==2.6.3 ; python_version >= "3.9" and python_version < "4.0"
-pygithub==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
+pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0"
+pygithub==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
 pygments==2.18.0 ; python_version >= "3.9" and python_version < "4.0"
 pyjwt[crypto]==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
 pynacl==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
 requests==2.32.3 ; python_version >= "3.9" and python_version < "4.0"
 rich==13.7.1 ; python_version >= "3.9" and python_version < "4.0"
-rtoml==0.10.0 ; python_version >= "3.9" and python_version < "4.0"
-typer==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
-typing-extensions==4.10.0 ; python_version >= "3.9" and python_version < "4.0"
+rtoml==0.11.0 ; python_version >= "3.9" and python_version < "4.0"
+shellingham==1.5.4 ; python_version >= "3.9" and python_version < "4.0"
+typer==0.12.3 ; python_version >= "3.9" and python_version < "4.0"
+typing-extensions==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
 urllib3==2.2.2 ; python_version >= "3.9" and python_version < "4.0"
 wrapt==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
->>>>>>> b4c9ea3 (Bump certifi from 2024.2.2 to 2024.7.4 (#213))
